requirements.txt
fastapi[all] >=0.100.0, <0.101.0
honcho >=1.1.0, <1.2.0
psycopg2--binary >=2.9.5, <2.10.0
pytest >=7.2.1, <7.3.0
pytest-cov >=4.1.0, <4.2.0
python-dotenv >=1.0.0, <1.1.0
requests >=2.31.0, <2.32.0
sqlalchemy >=2.0.4, <2.1.0
alembic >=1.10.2, <1.11.0
pygithub >=1.58.0, <1.59.0
black >=23.10.1, <23.11.0
pyjwt >=2.6.0, <2.7.0
python-jose[cryptography]
passlib[bcrypt]
email-validator